Final Project
This is what I have at the due date. It is still a work in progress but roughly 85% done. Note: some of the lines of code are //'d out for the presentation because they work but not as intended and would hinder me from showing other functions of the game.